poc/requirements.txt
aws-cdk-lib==2.72.1
constructs>=10.0.0,<11.0.0
aws_analytics_reference_architecture==2.10.0
aws-cdk.lambda-layer-kubectl-v22==2.0.7